third_party/blink/web_tests/external/wpt/fetch/metadata/generated/element-img.https.sub-expected.txt

This is a testharness.js-based test.
[FAIL] sec-fetch-site - src - HTTPS downgrade-upgrade, no attributes
  promise_test: Unhandled rejection with value: object "Error: Failed to query for recorded headers."
[FAIL] sec-fetch-storage-access - src - Cross-site, no attributes
  assert_array_equals: expected property 0 to be "none" but got "active" (expected array ["none"] got ["active"])
[FAIL] sec-fetch-storage-access - srcset - Cross-site, no attributes
  assert_array_equals: expected property 0 to be "none" but got "active" (expected array ["none"] got ["active"])
Harness: the test ran to completion.
